Refactor: Move ChatChannelNotificationType to bisq.chat.notifications
