+++ chat/src/main/java/bisq/chat/notifications/ChatChannelNotificationType.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.chat;
+package bisq.chat.notifications;
 
 import bisq.common.proto.ProtoEnum;
 import bisq.common.proto.ProtobufUtils;
